[components] Closes https://github.com/mozilla-mobile/android-components/issues/2421: Have getListenArticleMetadata take Long article ID.

+++ mobile/android/android-components/components/service/pocket/src/main/java/mozilla/components/service/pocket/PocketListenEndpoint.kt
@@ -30,7 +30,7 @@ class PocketListenEndpoint internal constructor(
      * [PocketResponse.Failure].
      */
     @WorkerThread // Synchronous network call.
-    fun getListenArticleMetadata(articleID: Int, articleURL: String): PocketResponse<PocketListenArticleMetadata> {
+    fun getListenArticleMetadata(articleID: Long, articleURL: String): PocketResponse<PocketListenArticleMetadata> {
         val json = rawEndpoint.getArticleListenMetadata(articleID, articleURL)
         val metadata = json?.let { jsonParser.jsonToListenArticleMetadata(it) }
         return PocketResponse.wrap(metadata)
